Updated main.py
On deleting any post, serial no was not changing for other posts and a gap of that serial no appears in our dashboard. Suppose we have posts with Sno 1,2,3,4,5,6,7,8,9,10 and after deleting post with serial no 7 the other posts appear as 1,2,3,4,5,6,   8,9,10.  
So i defined a function which will execute a query which will decrease the serial no of other posts which has serial no greater than that initially.
And i called that function inside delete function such that it will be updated every time when we will delete.
To see the effects, serial no should be in proper order (continious starting from 1) and also ensure that for table "post" in database in "Operations" tab the value of AUTO_INCREMENT should be set as one more than serial no of last row.